Prevent crash when starting an exercise
If an exercise is started with incorrect permissions or location
is disabled and it is being requested, WHS throws an exception,
this was not being handled in HealthServicesRepository which
resulted in a crash when calling startExercise. This change
catches the error and displays it in an error screen.

Implemented in ExerciseScreen to avoid having to reimplement the
error catching logic since starting an exercise is called from
multiple places.

Prevents crashes like the one seen in #212.